backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
playwright  
python-dotenv
pillow
google-generativeai